requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
python-multipart>=0.0.9
Pillow>=10.2.0
pillow-heif>=0.15.0
